Add InputActionRegistry and IInputActionResolver for action name mapping
- Introduced InputActionRegistry for stable InputActionId resolution.
- Implemented IInputActionResolver interface for action resolution.
- Updated InputBindingCompiler and related components to utilize the new action registry.
- Enhanced error handling and diagnostics for action name validation.

+++ engine/kettle/include/input/InputBindingTable.h
@@ -49,9 +49,10 @@ static constexpr uint16_t MaxMouseButtons = 8;
 
 struct InputBindingTable
 {
-	// Action name registry (indexed by ActionId.Value - 1, for debug/tooling)
+	// Action names from the action registry, indexed by ActionId.Value - 1.
+	// Debug/tooling only; runtime logic should compare InputActionIds.
 	std::vector<std::string> ActionNames;
-	uint16_t ActionCount = 0;
+	uint8_t ActionCount = 0;
 
 	// Keyboard: indexed by control code
 	std::array<BindingSlot, MaxKeyboardControls> KeyboardSlots{};
